Refactor, move class LocalBeanCache to PBEMSetupPanel -  PBEMSetupPanel.java is the only class that uses the 'LocalBeanCache.' The benefit here is we are reducing the overall complexity of the code base by isolating the 'LocalBeanCache' its one usage.
